+++ .storybook/config.js
@@ -1,12 +1,11 @@
 import { configure, addDecorator } from '@storybook/react';
-
+import './InstallStyles.js';
 import React from 'react';
 import ThemeProvider from '../src/ThemeProvider';
 import { setDefaults } from '@storybook/addon-info';
 import { withKnobs } from '@storybook/addon-knobs';
 import { setOptions } from '@storybook/addon-options';
 import { withInfo } from '@storybook/addon-info';
-
 
 import { withSmartKnobs } from 'storybook-addon-smart-knobs';