+++ babel.config.js
@@ -1,7 +1,7 @@
 module.exports = function presets(api) {
   const env = api.cache(() => process.env.NODE_ENV);
   const envConfig = { modules: false };
-  const plugins = [];
+  const plugins = ['inline-react-svg'];
   if (env === 'test') {
     delete envConfig.modules;
     plugins.push('@babel/plugin-transform-modules-commonjs');