+++ src/Menu/Drawer/Drawer.js
@@ -3,7 +3,7 @@ import { Drawer as MuiDrawer } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import { styled } from '@material-ui/styles';
 import { ListHeader } from './ListHeader';
-import { searchLogic } from './searchLogic';
+import { normalizeFunctionItems, searchLogic } from './searchLogic';
 import { SearchResultListing } from './SearchResultListing';
 import { MenuList } from './MenuList';
 
@@ -16,7 +16,15 @@ const StyledDiv = styled('div')({
   },
 });
 
-export const Drawer = ({ items, open, onClose }) => {
+export const Drawer = ({
+  items: oldItems,
+  open,
+  onClose,
+  title,
+  subtitle,
+  productName,
+}) => {
+  const items = normalizeFunctionItems(oldItems);
   const [search, setSearch] = useState('');
   let searchResults = [];
   if (search != null) {
@@ -26,7 +34,13 @@ export const Drawer = ({ items, open, onClose }) => {
     <MuiDrawer open={open} onClose={onClose}>
       <StyledDiv>
         <div>
-          <ListHeader search={search} setSearch={setSearch} />
+          <ListHeader
+            search={search}
+            setSearch={setSearch}
+            title={title}
+            subtitle={subtitle}
+            productName={productName}
+          />
         </div>
         <div>
           {!search && <MenuList closeDrawer={onClose} items={items} />}
@@ -47,7 +61,7 @@ export const Drawer = ({ items, open, onClose }) => {
 
 Drawer.defaultProps = {};
 const menuItemShape = PropTypes.shape({
-  title: PropTypes.string.isRequired,
+  title: PropTypes.oneOfType([PropTypes.string, PropTypes.func]).isRequired,
   component: PropTypes.object,
   componentProps: PropTypes.object,
 });
@@ -56,6 +70,9 @@ menuItemShape.children = PropTypes.arrayOf(PropTypes.shape(menuItemShape));
 Drawer.propTypes = {
   open: PropTypes.bool.isRequired,
   onClose: PropTypes.func.isRequired,
+  title: PropTypes.string.isRequired,
+  subtitle: PropTypes.string.isRequired,
+  productName: PropTypes.string.isRequired,
   items: PropTypes.arrayOf(menuItemShape).isRequired,
 };
 export default Drawer;