+++ src/globalStyle.js
@@ -1,23 +1,5 @@
-import { defaultGreen, defaultOrange, defaultRed, defaultGrey } from './colors';
-
 export const globalStyle = {
   marginTop: {
     margin: 10,
-  },
-  disabled: {
-    backgroundColor: defaultGrey,
-    color: 'white',
-  },
-  buttonColorSuccess: {
-    backgroundColor: defaultGreen,
-    color: 'white',
-  },
-  buttonColorSecondary: {
-    backgroundColor: defaultOrange,
-    color: 'white',
-  },
-  buttonColorThird: {
-    backgroundColor: defaultRed,
-    color: 'white',
   },
 };